+++ layout/app/home/index.js
@@ -6,7 +6,6 @@ import LayoutHome from './component';
 export default connect(
   (state) => ({
     responsive: state.responsive,
-    dashFeatured: state.dashboards.featured.list,
   }),
   null,
 )(LayoutHome);
