+++ pages/_app.jsx
@@ -13,7 +13,6 @@ import {
   setUser,
 } from 'redactions/user';
 import { setMobileDetect, mobileParser } from 'react-responsive-redux';
-import { getFeaturedDashboards } from 'modules/dashboards/actions';
 import { getPublishedPartners } from 'modules/partners/actions';
 import { setHostname } from 'redactions/common';
 
@@ -22,7 +21,6 @@ import { containsString } from 'utils/string';
 
 // constants
 import {
-  PAGES_WITHOUT_DASHBOARDS,
   FULLSCREEN_PAGES,
 } from 'constants/app';
 
@@ -52,16 +50,10 @@ class RWApp extends App {
     // sets user data coming from a request (server) or the store (client)
     const { user } = isServer ? req : store.getState();
     const {
-      dashboards: { featured: { list: featuredDashboards } },
       partners: { published: { list: publishedPartners } },
     } = store.getState();
     if (user) store.dispatch(setUser(user));
 
-    // fetches published featured dashboards
-    // to populate dashboards menu in the app header and footer
-    if (!containsString(pathname, PAGES_WITHOUT_DASHBOARDS) && !featuredDashboards.length) {
-      await store.dispatch(getFeaturedDashboards());
-    }
     // fetches partners for footer
     if (!containsString(pathname, FULLSCREEN_PAGES) && !publishedPartners.length) {
       await store.dispatch(getPublishedPartners());
